minor updates to address warnings raised by current GCC compilers, tested with GCC 13
